requirements.txt
ccxt==4.4.64
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.43.2
yfinance==0.2.54
